Google-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
certifi==2026.2.25
    # via requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==46.0.6
    # via google-auth
fastapi==0.135.3
    # via -r requirements.in
google-api-core==2.30.2
    # via google-api-python-client
google-api-python-client==2.193.0
    # via -r requirements.in
google-auth==2.49.1
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.1
    # via google-api-python-client
google-auth-oauthlib==1.3.1
    # via -r requirements.in
googleapis-common-protos==1.74.0
    # via google-api-core
h11==0.16.0
    # via uvicorn
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
oauthlib==3.3.1
    # via requests-oauthlib
proto-plus==1.27.2
    # via google-api-core
protobuf==7.34.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pyparsing==3.3.2
    # via httplib2
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.3
    # via uvicorn
requests==2.33.1
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
tzdata==2026.1
    # via -r requirements.in
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.43.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
